+++ js/utils/extensions.js
@@ -77,36 +77,62 @@ String.prototype.contains = function() {
 
 String.prototype.toDate = function(format)
 {
-    var normalized      = this.replace(/[^a-zA-Z0-9%]/g, '-');
-    var normalizedFormat= format.toLowerCase().replace(/[^a-zA-Z0-9%]/g, '-');
-    var formatItems     = normalizedFormat.split('-');
-    var dateItems       = normalized.split('-');
+    var normalized, normalizedFormat, formatItems, dateItems;
+    var monthIndex, dayIndex, yearIndex, hourIndex, minutesIndex, secondsIndex;
+    var today, year, month, day, hour, minute, second;
 
-    var monthIndex  = formatItems.indexOf("mm");
-    var dayIndex    = formatItems.indexOf("dd");
-    var yearIndex   = formatItems.indexOf("yyyy") > -1 ? formatItems.indexOf("yyyy") : formatItems.indexOf("yy");
-    var hourIndex     = formatItems.indexOf("hh");
-    var minutesIndex  = formatItems.indexOf("ii");
-    var secondsIndex  = formatItems.indexOf("ss");
+    if (!Utils.isValue(format)) {
+        format = "yyyy-mm-dd";
+    }
 
-    var monthIndexP  = formatItems.indexOf("%m");
-    var dayIndexP    = formatItems.indexOf("%d");
-    var yearIndexP   = formatItems.indexOf("%y") > -1 ? formatItems.indexOf("%y") : formatItems.indexOf("%Y");
-    var hourIndexP     = formatItems.indexOf("%h");
-    var minutesIndexP  = formatItems.indexOf("%i");
-    var secondsIndexP  = formatItems.indexOf("%s");
+    normalized      = this.replace(/[^a-zA-Z0-9%]/g, '-');
+    normalizedFormat= format.toLowerCase().replace(/[^a-zA-Z0-9%]/g, '-');
+    formatItems     = normalizedFormat.split('-');
+    dateItems       = normalized.split('-');
 
-    var today = new Date();
+    monthIndex  = formatItems.indexOf("mm") > -1 ? formatItems.indexOf("mm") : formatItems.indexOf("%m");
+    dayIndex    = formatItems.indexOf("dd") > -1 ? formatItems.indexOf("dd") : formatItems.indexOf("%d");
+    yearIndex   = formatItems.indexOf("yyyy") > -1 ? formatItems.indexOf("yyyy") : formatItems.indexOf("yy") > -1 ? formatItems.indexOf("yy") : formatItems.indexOf("%y");
+    hourIndex     = formatItems.indexOf("hh") > -1 ? formatItems.indexOf("hh") : formatItems.indexOf("%h");
+    minutesIndex  = formatItems.indexOf("ii") > -1 ? formatItems.indexOf("ii") : formatItems.indexOf("mi") > -1 ? formatItems.indexOf("mi") : formatItems.indexOf("%i");
+    secondsIndex  = formatItems.indexOf("ss") > -1 ? formatItems.indexOf("ss") : formatItems.indexOf("%s");
 
-    var year  = yearIndex >-1 ? dateItems[yearIndex] : yearIndexP > -1 ? dateItems[yearIndexP] : today.getFullYear();
-    var month = monthIndex >-1 ? dateItems[monthIndex]-1 : monthIndexP ? dateItems[monthIndexP] - 1 : today.getMonth()-1;
-    var day   = dayIndex >-1 ? dateItems[dayIndex] : dayIndexP > -1 ? dateItems[dayIndexP] : today.getDate();
+    today = new Date();
 
-    var hour    = hourIndex >-1 ? dateItems[hourIndex] : hourIndexP > -1 ? dateItems[hourIndexP] : today.getHours();
-    var minute  = minutesIndex>-1 ? dateItems[minutesIndex] : minutesIndexP > -1 ? dateItems[minutesIndexP] : today.getMinutes();
-    var second  = secondsIndex>-1 ? dateItems[secondsIndex] : secondsIndexP > -1 ? dateItems[secondsIndexP] : today.getSeconds();
+    year  = yearIndex >-1 ? dateItems[yearIndex] : today.getFullYear();
+    month = monthIndex >-1 ? dateItems[monthIndex]-1 : today.getMonth()-1;
+    day   = dayIndex >-1 ? dateItems[dayIndex] : today.getDate();
+
+    hour    = hourIndex >-1 ? dateItems[hourIndex] : today.getHours();
+    minute  = minutesIndex>-1 ? dateItems[minutesIndex] : today.getMinutes();
+    second  = secondsIndex>-1 ? dateItems[secondsIndex] : today.getSeconds();
 
     return new Date(year,month,day,hour,minute,second);
+};
+
+Date.prototype.getWeek = function (dowOffset) {
+    var nYear, nday, newYear, day, daynum, weeknum;
+
+    dowOffset = !Utils.isValue(dowOffset) ? METRO_WEEK_START : typeof dowOffset === 'number' ? parseInt(dowOffset) : 0;
+    newYear = new Date(this.getFullYear(),0,1);
+    day = newYear.getDay() - dowOffset;
+    day = (day >= 0 ? day : day + 7);
+    daynum = Math.floor((this.getTime() - newYear.getTime() -
+        (this.getTimezoneOffset()-newYear.getTimezoneOffset())*60000)/86400000) + 1;
+
+    if(day < 4) {
+        weeknum = Math.floor((daynum+day-1)/7) + 1;
+        if(weeknum > 52) {
+            nYear = new Date(this.getFullYear() + 1,0,1);
+            nday = nYear.getDay() - dowOffset;
+            nday = nday >= 0 ? nday : nday + 7;
+            weeknum = nday < 4 ? 1 : 53;
+        }
+    }
+    else {
+        weeknum = Math.floor((daynum+day-1)/7);
+    }
+    return weeknum;
 };
 
 Date.prototype.getYear = function(){
@@ -150,7 +176,7 @@ Date.prototype.format = function(format, locale){
             '%c': date.toUTCString(),
             '%C': Math.floor(nYear/100),
             '%d': zeroPad(nDate, 2),
-            // 'dd': zeroPad(nDate, 2),
+            'dd': zeroPad(nDate, 2),
             '%e': nDate,
             '%F': date.toISOString().slice(0,10),
             '%G': getThursday().getFullYear(),